consolidate ept:: and copc:: Connector classes into a single one
